requirements.txt
streamlit
pandas
python-docx
openpyxl
PyPDF2
platform


